make btree the only inner prefilter
