+++ Cargo.toml
@@ -5,10 +5,7 @@ edition = "2024"
 
 [dependencies]
 regex-syntax = "0.8"
-regex-automata = "0.4"
-aho-corasick = "1.1"
 roaring = "0.11"
-fst = "0.4"
 
 [dev-dependencies]
 criterion = { version = "0.8.1", features = ["html_reports"] }
